Checkpoint commit. Interfaces aren't passing tests
